Header (not finished)

+++ 3.F_BUY.c
@@ -54,7 +54,7 @@ void readProduct(struct product **products, int *productCount) {
     }
 }
 
-//show product/;llkkkkkkkkk jkllkl
+//show product
 void showProduct(struct product *Arrproduct, int number)
 {
 	for (int i = 0; i < number; i++)
@@ -68,15 +68,14 @@ void showProduct(struct product *Arrproduct, int number)
 
 
 // Function to add new cart
-void addCart(struct cart **carts, int *cartCount, char *product_id, char *product_name, float price, float sum, char *coupon, float final_sum, int user_id)
+void addCart(struct cart **carts, int *cartCount, char *product_id, char *product_name, int price, int sum, char *coupon, int final_sum, int user_id)
 {
     *carts = (struct cart *)realloc(*carts, (*cartCount + 1) * sizeof(struct cart));
-    (*carts)[*cartCount].id = ++lastAssignedID; // Increment and assign new ID
+    (*carts)[*cartCount].id = ++lastAssignedID;
     strcpy((*carts)[*cartCount].product_id, product_id);
     strcpy((*carts)[*cartCount].product_name, product_name);
     (*carts)[*cartCount].price = price;
     (*carts)[*cartCount].sum = sum;
-    //strcpy((*carts)[*cartCount].coupon, coupon);
     (*carts)[*cartCount].final_sum = final_sum;
     (*carts)[*cartCount].user_id = user_id;
     (*cartCount)++;
